Async api code and solving quiz

+++ api/Models/Dto/ScoreDto/AddScoreDto.cs
@@ -2,6 +2,6 @@
 
 public class AddScoreDto
 {
-    public int PercentOfCorrectAnswers { get; set; }
+    public decimal PercentOfCorrectAnswers { get; set; }
     public int QuizId { get; set; }
 }
