+++ api/Models/Dto/ScoreDto/GetScoreDto.cs
@@ -4,7 +4,7 @@ public class GetScoreDto
 {
     public int Id { get; set; }
     public DateTime DateOfCompletion { get; set; }
-    public int PercentOfCorrectAnswers { get; set; }
+    public decimal PercentOfCorrectAnswers { get; set; }
     public int QuizId { get; set; }
     public int UserId { get; set; }
 }
